fix: 🐞 fix toast newsletter form
Move footer newsletter tests into newsletter.spec.ts and tag Mailchimp
integration tests with @manual.
