--- apps/cookbook/playwright-manual.config.ts
@@ -1,6 +1,0 @@
-import { defineConfig } from '@playwright/test';
-import baseConfig from './playwright.config';
-
-export default defineConfig(baseConfig, {
-  testMatch: /.*\.manual\.ts/,
-});
